fix: clicking on an already clicked subtitle not working
If the user clicks on a subtitle in the tracklist to jump at that point
in time twice, then the second time will not work. We fix this by
updating current time in the Store by updating it every time the video
player updates its time. In addition, it is necessary to make a
distinction where the time update is coming from, either the VideoPlayer
or the SubtitleTracklistViewer because since the VideoPlayer is
listening for updates on this time variable, it will stutter every time
the VideoPlayer makes an update to the current time in the Store. Hence,
we only want to listen when we get a TimeEvent from the
SubtitleTracklistViewer.